Reworked sidenav, status exposed over a service, closes on click
The navigation component now completely contains the material sidenav component and manages its state via service. That allows it to be controlled from everywhere within the application and also to close on navigation click. In the future this behaviour could be extended to only close on successful navigation transition but this is okay for now.